requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
Flask-Cors==3.0.10
fsspec==2023.4.0
fugashi==1.2.1
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.2
lit==16.0.2
loguru==0.7.0
manga-ocr==0.1.10
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==9.5.0
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.0.9
pyperclip==1.8.2
pytesseract==0.3.10
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
six==1.16.0
sympy==1.11.1
termcolor==2.3.0
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
unidic-lite==1.0.8
urllib3==1.26.15
Werkzeug==2.3.3
